code improvement duplicates Signed-off-by: dbulahov <bulahovdenis@gmail.com>
