+++ xmla/ws/jakarta.model.xmla/src/main/java/org/eclipse/daanse/xmla/ws/jakarta/model/xmla/xmla/AttributeRelationship.java
@@ -37,7 +37,7 @@ public class AttributeRelationship {
     @XmlElement(name = "OverrideBehavior")
     protected String overrideBehavior;
     @XmlElement(name = "Annotations")
-    protected AttributeRelationship.Annotations annotations;
+    protected Annotations annotations;
     @XmlElement(name = "Name")
     protected String name;
     @XmlElement(name = "Visible")
@@ -85,11 +85,11 @@ public class AttributeRelationship {
         this.overrideBehavior = value;
     }
 
-    public AttributeRelationship.Annotations getAnnotations() {
+    public Annotations getAnnotations() {
         return annotations;
     }
 
-    public void setAnnotations(AttributeRelationship.Annotations value) {
+    public void setAnnotations(Annotations value) {
         this.annotations = value;
     }
 
@@ -115,22 +115,6 @@ public class AttributeRelationship {
 
     public void setTranslations(AttributeRelationship.Translations value) {
         this.translations = value;
-    }
-
-    @XmlAccessorType(XmlAccessType.FIELD)
-    @XmlType(name = "", propOrder = {"annotation"})
-    public static class Annotations {
-
-        @XmlElement(name = "Annotation")
-        protected List<Annotation> annotation;
-
-        public List<Annotation> getAnnotation() {
-            return this.annotation;
-        }
-
-        public void setAnnotation(List<Annotation> annotation) {
-            this.annotation = annotation;
-        }
     }
 
     @XmlAccessorType(XmlAccessType.FIELD)
