+++ xmla/ws/jakarta.model.xmla/src/main/java/org/eclipse/daanse/xmla/ws/jakarta/model/xmla/xmla/Level.java
@@ -39,7 +39,7 @@ public class Level {
     @XmlElement(name = "Translations")
     protected Level.Translations translations;
     @XmlElement(name = "Annotations")
-    protected Level.Annotations annotations;
+    protected Annotations annotations;
 
     public String getName() {
         return name;
@@ -89,28 +89,12 @@ public class Level {
         this.translations = value;
     }
 
-    public Level.Annotations getAnnotations() {
+    public Annotations getAnnotations() {
         return annotations;
     }
 
-    public void setAnnotations(Level.Annotations value) {
+    public void setAnnotations(Annotations value) {
         this.annotations = value;
-    }
-
-    @XmlAccessorType(XmlAccessType.FIELD)
-    @XmlType(name = "", propOrder = {"annotation"})
-    public static class Annotations {
-
-        @XmlElement(name = "Annotation")
-        protected List<Annotation> annotation;
-
-        public List<Annotation> getAnnotation() {
-            return this.annotation;
-        }
-
-        public void setAnnotation(List<Annotation> annotation) {
-            this.annotation = annotation;
-        }
     }
 
     @XmlAccessorType(XmlAccessType.FIELD)
