+++ xmla/ws/jakarta.model.xmla/src/main/java/org/eclipse/daanse/xmla/ws/jakarta/model/xmla/xmla/MiningModelColumn.java
@@ -45,7 +45,7 @@ public class MiningModelColumn {
     @XmlElement(name = "ModelingFlags")
     protected MiningModelColumn.ModelingFlags modelingFlags;
     @XmlElement(name = "Annotations")
-    protected MiningModelColumn.Annotations annotations;
+    protected Annotations annotations;
 
     public String getName() {
         return name;
@@ -119,28 +119,12 @@ public class MiningModelColumn {
         this.modelingFlags = value;
     }
 
-    public MiningModelColumn.Annotations getAnnotations() {
+    public Annotations getAnnotations() {
         return annotations;
     }
 
-    public void setAnnotations(MiningModelColumn.Annotations value) {
+    public void setAnnotations(Annotations value) {
         this.annotations = value;
-    }
-
-    @XmlAccessorType(XmlAccessType.FIELD)
-    @XmlType(name = "", propOrder = {"annotation"})
-    public static class Annotations {
-
-        @XmlElement(name = "Annotation")
-        protected List<Annotation> annotation;
-
-        public List<Annotation> getAnnotation() {
-            return this.annotation;
-        }
-
-        public void setAnnotation(List<Annotation> annotation) {
-            this.annotation = annotation;
-        }
     }
 
     @XmlAccessorType(XmlAccessType.FIELD)
